+++ src/comment/controller.ts
@@ -6,14 +6,18 @@ import {
   DeleteComment,
 } from './service';
 import { successResponse } from '../utils/response';
+import { UnauthorizedError } from '../utils/errors';
 
 export const createCommentController = async (
   req: Request,
   res: Response,
   next: NextFunction
 ) => {
+  if (!req.user) {
+    throw new UnauthorizedError('You must be logged in to perform this action');
+  }
   try {
-    const user = (req as any).user;
+    const user = req.user;
     const commentData = req.body;
 
     const comment = await CreateComment(user, commentData);
@@ -30,8 +34,11 @@ export const getCommentByIdController = async (
   res: Response,
   next: NextFunction
 ) => {
+  if (!req.user) {
+    throw new UnauthorizedError('You must be logged in to perform this action');
+  }
   try {
-    const user = (req as any).user;
+    const user = req.user;
     const commentId = req.params.id;
 
     const comment = await GetCommentById(user, commentId);
@@ -48,8 +55,11 @@ export const getCommentsByPostIdController = async (
   res: Response,
   next: NextFunction
 ) => {
+  if (!req.user) {
+    throw new UnauthorizedError('You must be logged in to perform this action');
+  }
   try {
-    const user = (req as any).user;
+    const user = req.user;
     const postId = req.params.id;
 
     const comments = await GetCommentsByPostId(user, postId);
@@ -66,8 +76,11 @@ export const deleteCommentController = async (
   res: Response,
   next: NextFunction
 ) => {
+  if (!req.user) {
+    throw new UnauthorizedError('You must be logged in to perform this action');
+  }
   try {
-    const user = (req as any).user;
+    const user = req.user;
     const commentId = req.params.id;
 
     await DeleteComment(user, commentId);
